baskervville is the way it was spelled on the google font website and it is working because if i remove it and the sans serif i get a diffrent font.
i do have the a picture id in the html its for the logo image specificly and it deals with the scale and background color of the image the logo id deals with the placement and border of the title and was named incorrectly.

header top nav were leftovers from the lawfirm code that i forgot to remove.

i changed the accent color.